Refactored App.vue into DrinkList #5
DrinkCards are now held inside the new DrinkList component. Also renamed header and footer, and removed an unused function that was hiding in App.vue
