+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "core-js": "^3.8.3",
     "vue": "^3.2.13",
+    "vue-content-loader": "^2.0.1",
     "vue-horizontal": "^1.0.2"
   },
   "devDependencies": {
